Module clipboard_win::raw

Raw bindings to Windows clipboard.

General information

All pre & post conditions are stated in description of functions.

Open clipboard

To access any information inside clipboard it is necessary to open it by means of open().

After that Clipboard cannot be opened any more until close() is called.

Structs

EnumFormats
Enumerator over available clipboard formats.

Functions

close
Closes clipboard.

count_formats
Retrieves number of currently available formats on clipboard.

empty
Empties clipboard.

format_name
Returns format name based on it's code.

get
Retrieves data of specified format from clipboard.

get_clipboard_data
Retrieves raw pointer to clipboard data.

get_file_list
Retrieves a list of file paths from the CF_HDROP format.

get_string
Retrieves String from CF_UNICODETEXT format

is_format_avail
Determines whenever provided clipboard format is available on clipboard or not.

open
Opens clipboard.

register_format
Registers a new clipboard format with specified name.

seq_num
Retrieves clipboard sequence number.

set
Sets data onto clipboard with specified format.

size
Retrieves size of clipboard data for specified format.

size_unsafe⚠
Retrieves size of clipboard data for specified format.