Struct clipboard_win::ClipboardManager

pub struct ClipboardManager {
    // some fields omitted
}

Clipboard manager provides a primitive hack for console application to handle updates of clipboard. It uses get_clipboard_seq_num to determines whatever clipboard is updated or not. Due to that this manager is a bit hacky and not exactly right way to properly work with clipboard. On other hand it is the best and most easy option for console application as a proper window is required to be created to work with clipboard.

Methods

impl ClipboardManager

fn new() -> ClipboardManager

Constructs new ClipboardManager with default settings

fn delay(&mut self, delay_ms: u32) -> &mut ClipboardManager

Configure manager's delay between accessing clipboard.

fn ok_callback(&mut self, callback: fn(&String)) -> &mut ClipboardManager

Sets callback for successfully retrieved clipboard's text.

fn err_callback(&mut self, callback: fn(&WindowsError)) -> &mut ClipboardManager

Sets callback for failed retrieval of clipboard's text.

Error code is passed from get_clipboard_string()

fn run(&self)

Starts manager loop.

It's infinitely running with delay and checking whatever clipboard is updated. In case if it is updated callbacks will be called. Depending on whatever clipboard's text was retrieved or not right callback will be called.