Struct clipboard_win::WinError

pub struct WinError {
    // some fields omitted
}

Represents Windows error.

result indicates whatever command was executed successfully or not.

errno extended Windows error code. Error codes

Methods

impl WinError

fn new(result: bool, errno: u32) -> WinError

Constructs new error.

fn is_ok(&self) -> bool

Returns true if result is ok

fn is_not_ok(&self) -> bool

Returns true if result is not ok

fn errno(&self) -> u32

Returns extended error code. Should be used in case if result is not ok.

fn errno_desc(&self) -> String

Returns description of WinAPI error code.