clipboard_win
Trait Getter

pub trait Getter<Type> {
    // Required method
    fn read_clipboard(&self, out: &mut Type) -> SysResult<usize>;
}

Describes format getter, specifying data type as type param

Default implementations only perform write, without opening/closing clipboard

Required Methods

fn read_clipboard(&self, out: &mut Type) -> SysResult<usize>

Reads content of clipboard into out, returning number of bytes read on success, or otherwise 0.