clipboard_win
Trait Setter

pub trait Setter<Type: ?Sized> {
    // Required method
    fn write_clipboard(&self, data: &Type) -> SysResult<()>;
}

Describes format setter, specifying data type as type param

Default implementations only perform write, without opening/closing clipboard

Required Methods

fn write_clipboard(&self, data: &Type) -> SysResult<()>

Writes content of data onto clipboard, returning whether it was successful or not