Module clipboard_win::wrapper

Provides direct wrappers to WinAPI functions.

These functions omit calls to OpenClipboard and CloseClipboard. Due to that most functions have requirements for them to be called.

Functions

close_clipboard
Wrapper around CloseClipboard.

count_formats
Retrieves number of currently available formats on clipboard.

empty_clipboard
Wrapper around EmptyClipboard.

get_clipboard
Wrapper around GetClipboardData.

get_clipboard_formats
Extracts available clipboard formats.

get_clipboard_seq_num
Wrapper around GetClipboardSequenceNumber.

get_clipboard_string
Wrapper around GetClipboardData with hardcoded UTF16 format.

get_format_name
Returns format name based on it's code.

get_last_error
Wrapper around GetLastError.

is_format_avail
Determines whenever provided clipboard format is available on clipboard or not.

open_clipboard
Wrapper around OpenClipboard.

register_format
Registers a new clipboard format with specified name.

set_clipboard
Wrapper around SetClipboardData to set unicode text.

set_clipboard_raw
Wrapper around SetClipboardData to set raw data.