Crate clipboard_win

Clipboard WinAPI

This crate provide simple means to operate with Windows clipboard.

Example:

extern crate clipboard_win;

use clipboard_win::*;

fn main() {
    println!("I set some clipboard text like a boss!");
    set_clipboard("for my waifu!");
}

Modules

wrapper
Provides direct wrappers to WinAPI functions.

Structs

ClipboardManager
Clipboard manager provides a primitive hack for console application to handle updates of clipboard. It uses get_clipboard_seq_num to determines whatever clipboard is updated or not. Due to that this manager is a bit hacky and not exactly right way to properly work with clipboard. On other hand it is the best and most easy option for console application as a proper window is required to be created to work with clipboard.

WindowsError
Represents Windows error code.

Functions

get_clipboard
Retrieves clipboard content.

get_clipboard_formats
Extracts available clipboard formats.

get_clipboard_string
Retrieves clipboard content in UTF16 format and convert it to String.

get_format_name
Returns format name based on it's code.

set_clipboard
Set clipboard with text.