clipboard_win
Module formats

Standard clipboard formats.

Header: Winuser.h

Description is taken from Standard Clipboard Formats

Structs

Bitmap

Format for bitmap images i.e. CF_BITMAP.

FileList

Format for file lists (generated by drag & drop).

Html

HTML Foramt

RawData

Format to write/read from clipboard as raw bytes

Unicode

Format to read/write unicode string.

Constants

CF_BITMAP

A handle to a bitmap (HBITMAP).

CF_DIB

A memory object containing a BITMAPINFO structure followed by the bitmap bits.

CF_DIBV5

A memory object containing a BITMAPV5HEADER structure followed by the bitmap color space information and the bitmap bits.

CF_DIF

Software Arts’ Data Interchange Format.

CF_DSPBITMAP

Bitmap display format associated with a private format. The hMem parameter must be a handle to data that can be displayed in bitmap format in lieu of the privately formatted data.

CF_DSPENHMETAFILE

Enhanced metafile display format associated with a private format. The hMem parameter must be a handle to data that can be displayed in enhanced metafile format in lieu of the privately formatted data.

CF_DSPMETAFILEPICT

Metafile-picture display format associated with a private format. The hMem parameter must be a handle to data that can be displayed in metafile-picture format in lieu of the privately formatted data.

CF_DSPTEXT

Text display format associated with a private format. The hMem parameter must be a handle to data that can be displayed in text format in lieu of the privately formatted data.

CF_ENHMETAFILE

A handle to an enhanced metafile (HENHMETAFILE).

CF_GDIOBJFIRST

Start of a range of integer values for application-defined GDI object clipboard formats.

CF_GDIOBJLAST

End of a range of integer values for application-defined GDI object clipboard formats.

CF_HDROP

A handle to type HDROP that identifies a list of files.

CF_LOCALE

The data is a handle to the locale identifier associated with text in the clipboard.

CF_METAFILEPICT

Handle to a metafile picture format as defined by the METAFILEPICT structure.

CF_OEMTEXT

Text format containing characters in the OEM character set.

CF_OWNERDISPLAY

Owner-display format.

CF_PALETTE

Handle to a color palette.

CF_PENDATA

Data for the pen extensions to the Microsoft Windows for Pen Computing.

CF_PRIVATEFIRST

Start of a range of integer values for private clipboard formats.

CF_PRIVATELAST

End of a range of integer values for private clipboard formats.

CF_RIFF

Represents audio data more complex than can be represented in a CF_WAVE standard wave format.

CF_SYLK

Microsoft Symbolic Link (SYLK) format.

CF_TEXT

ANSI text format.

CF_TIFF

Tagged-image file format.

CF_UNICODETEXT

UTF16 text format.

CF_WAVE

Represents audio data in one of the standard wave formats.

Traits

Format

Format trait